+++ pennylane_lightning/lightning_base.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 
 r"""
-This module contains the base class for all PennyLane Lightning simulator device.
-This devices interfaces with C++ for improved performance.
+This module contains the base class for all PennyLane Lightning simulator devices, and interfaces with C++ for improved performance.
 """
+
 from typing import List
 
 # from warnings import warn
@@ -53,9 +53,6 @@ try:
         # MeasuresC128,
         StateVectorC128,
         backend_info,
-        allocate_aligned_array,
-        get_alignment,
-        best_alignment,
     )
 
     from ._serialize import _serialize_ob, _serialize_observables, _serialize_ops
@@ -143,25 +140,6 @@ if CPP_BINARY_AVAILABLE:
                 returns_state=True,
             )
             return capabilities
-
-        @staticmethod
-        def _asarray(arr, dtype=None):
-            arr = np.asarray(arr)  # arr is not copied
-
-            if arr.dtype.kind not in ["f", "c"]:
-                return arr
-
-            if not dtype:
-                dtype = arr.dtype
-
-            # We allocate a new aligned memory and copy data to there if alignment or dtype mismatches
-            # Note that get_alignment does not necessarily return CPUMemoryModel(Unaligned) even for
-            # numpy allocated memory as the memory location happens to be aligned.
-            if int(get_alignment(arr)) < int(best_alignment()) or arr.dtype != dtype:
-                new_arr = allocate_aligned_array(arr.size, np.dtype(dtype)).reshape(arr.shape)
-                np.copyto(new_arr, arr)
-                arr = new_arr
-            return arr
 
         # To be able to validate the adjoint method [_validate_adjoint_method(device)],
         #  the qnode requires the definition of:
@@ -703,320 +681,6 @@ if CPP_BINARY_AVAILABLE:
         #         filter(lambda m: m.obs is None or not skip_diagonalizing(m.obs), circuit.measurements)
         #     )
         #     return super()._get_diagonalizing_gates(qml.tape.QuantumScript(measurements=meas_filtered))
-
-    if backend_info()["NAME"] == "lightning.qubit":
-
-        def _chunk_iterable(it, num_chunks):
-            "Lazy-evaluated chunking of given iterable from https://stackoverflow.com/a/22045226"
-            it = iter(it)
-            return iter(lambda: tuple(islice(it, num_chunks)), ())
-
-        allowed_operations = {
-            "Identity",
-            "BasisState",
-            "QubitStateVector",
-            "QubitUnitary",
-            "ControlledQubitUnitary",
-            "MultiControlledX",
-            "DiagonalQubitUnitary",
-            "PauliX",
-            "PauliY",
-            "PauliZ",
-            "MultiRZ",
-            "Hadamard",
-            "S",
-            "Adjoint(S)",
-            "T",
-            "Adjoint(T)",
-            "SX",
-            "Adjoint(SX)",
-            "CNOT",
-            "SWAP",
-            "ISWAP",
-            "PSWAP",
-            "Adjoint(ISWAP)",
-            "SISWAP",
-            "Adjoint(SISWAP)",
-            "SQISW",
-            "CSWAP",
-            "Toffoli",
-            "CY",
-            "CZ",
-            "PhaseShift",
-            "ControlledPhaseShift",
-            "CPhase",
-            "RX",
-            "RY",
-            "RZ",
-            "Rot",
-            "CRX",
-            "CRY",
-            "CRZ",
-            "CRot",
-            "IsingXX",
-            "IsingYY",
-            "IsingZZ",
-            "IsingXY",
-            "SingleExcitation",
-            "SingleExcitationPlus",
-            "SingleExcitationMinus",
-            "DoubleExcitation",
-            "DoubleExcitationPlus",
-            "DoubleExcitationMinus",
-            "QubitCarry",
-            "QubitSum",
-            "OrbitalRotation",
-            "QFT",
-            "ECR",
-        }
-
-        allowed_observables = {
-            "PauliX",
-            "PauliY",
-            "PauliZ",
-            "Hadamard",
-            "Hermitian",
-            "Identity",
-            "Projector",
-            "SparseHamiltonian",
-            "Hamiltonian",
-            "Sum",
-            "SProd",
-            "Prod",
-            "Exp",
-        }
-
-        class LightningQubit(LightningBase):
-            """PennyLane Lightning device.
-
-            A device that interfaces with C++ to perform fast linear algebra calculations.
-
-            Use of this device requires pre-built binaries or compilation from source. Check out the
-            :doc:`/installation` guide for more details.
-
-            Args:
-                wires (int): the number of wires to initialize the device with
-                c_dtype: Datatypes for statevector representation. Must be one of ``np.complex64`` or ``np.complex128``.
-                shots (int): How many times the circuit should be evaluated (or sampled) to estimate
-                    the expectation values. Defaults to ``None`` if not specified. Setting
-                    to ``None`` results in computing statistics like expectation values and
-                    variances analytically.
-                mcmc (bool): Determine whether to use the approximate Markov Chain Monte Carlo sampling method when generating samples.
-                kernel_name (str): name of transition kernel. The current version supports two kernels: ``"Local"`` and ``"NonZeroRandom"``.
-                    The local kernel conducts a bit-flip local transition between states. The local kernel generates a
-                    random qubit site and then generates a random number to determine the new bit at that qubit site. The ``"NonZeroRandom"`` kernel
-                    randomly transits between states that have nonzero probability.
-                num_burnin (int): number of steps that will be dropped. Increasing this value will
-                result in a closer approximation but increased runtime.
-                batch_obs (bool): Determine whether we process observables in parallel when computing the
-                    jacobian. This value is only relevant when the lightning qubit is built with OpenMP.
-            """
-
-            name = "Lightning Qubit PennyLane plugin"
-            short_name = "lightning.qubit"
-            operations = allowed_operations
-            observables = allowed_observables
-
-            def __init__(
-                self,
-                wires,
-                *,
-                c_dtype=np.complex128,
-                shots=None,
-                mcmc=False,
-                kernel_name="Local",
-                num_burnin=100,
-                batch_obs=False,
-                analytic=None,
-            ):
-                super().__init__(wires, shots=shots, c_dtype=c_dtype, analytic=analytic)
-
-                # Create the initial state. Internally, we store the
-                # state as an array of dimension [2]*wires.
-                self._state = self._create_basis_state(0)
-                self._pre_rotated_state = self._state
-
-                self._mcmc = mcmc
-                if self._mcmc:
-                    if kernel_name not in [
-                        "Local",
-                        "NonZeroRandom",
-                    ]:
-                        raise NotImplementedError(
-                            f"The {kernel_name} is not supported and currently only 'Local' and 'NonZeroRandom' kernels are supported."
-                        )
-                    if num_burnin >= shots:
-                        raise ValueError("Shots should be greater than num_burnin.")
-                    self._kernel_name = kernel_name
-                    self._num_burnin = num_burnin
-
-            def _create_basis_state(self, index):
-                """Return a computational basis state over all wires.
-                Args:
-                    index (int): integer representing the computational basis state
-                Returns:
-                    array[complex]: complex array of shape ``[2]*self.num_wires``
-                    representing the statevector of the basis state
-                Note: This function does not support broadcasted inputs yet.
-                """
-                state = np.zeros(2**self.num_wires, dtype=np.complex128)
-                state[index] = 1
-                state = self._asarray(state, dtype=self.C_DTYPE)
-                return self._reshape(state, [2] * self.num_wires)
-
-            def reset(self):
-                """Reset the device"""
-                super().reset()
-
-                # init the state vector to |00..0>
-                self._state = self._create_basis_state(0)
-                self._pre_rotated_state = self._state
-
-            @property
-            def state(self):
-                # Flattening the state.
-                shape = (1 << self.num_wires,)
-                return self._reshape(self._pre_rotated_state, shape)
-
-            def _apply_state_vector(self, state, device_wires):
-                """Initialize the internal state vector in a specified state.
-                Args:
-                    state (array[complex]): normalized input state of length ``2**len(wires)``
-                        or broadcasted state of shape ``(batch_size, 2**len(wires))``
-                    device_wires (Wires): wires that get initialized in the state
-                """
-
-                # translate to wire labels used by device
-                device_wires = self.map_wires(device_wires)
-
-                state = self._asarray(state, dtype=self.C_DTYPE)
-                output_shape = [2] * self.num_wires
-
-                if (
-                    len(device_wires) == self.num_wires
-                    and Wires(sorted(device_wires)) == device_wires
-                ):
-                    # Initialize the entire device state with the input state
-                    self._state = self._reshape(state, output_shape)
-                    return
-
-                # generate basis states on subset of qubits via the cartesian product
-                basis_states = np.array(list(product([0, 1], repeat=len(device_wires))))
-
-                # get basis states to alter on full set of qubits
-                unravelled_indices = np.zeros((2 ** len(device_wires), self.num_wires), dtype=int)
-                unravelled_indices[:, device_wires] = basis_states
-
-                # get indices for which the state is changed to input state vector elements
-                ravelled_indices = np.ravel_multi_index(unravelled_indices.T, [2] * self.num_wires)
-
-                state = self._scatter(ravelled_indices, state, [2**self.num_wires])
-                state = self._reshape(state, output_shape)
-                self._state = self._asarray(state, dtype=self.C_DTYPE)
-
-            def _apply_basis_state(self, state, wires):
-                """Initialize the state vector in a specified computational basis state.
-
-                Args:
-                    state (array[int]): computational basis state of shape ``(wires,)``
-                        consisting of 0s and 1s.
-                    wires (Wires): wires that the provided computational state should be initialized on
-
-                Note: This function does not support broadcasted inputs yet.
-                """
-                # translate to wire labels used by device
-                device_wires = self.map_wires(wires)
-
-                # length of basis state parameter
-                n_basis_state = len(state)
-
-                if not set(state.tolist()).issubset({0, 1}):
-                    raise ValueError("BasisState parameter must consist of 0 or 1 integers.")
-
-                if n_basis_state != len(device_wires):
-                    raise ValueError("BasisState parameter and wires must be of equal length.")
-
-                # get computational basis state number
-                basis_states = 2 ** (self.num_wires - 1 - np.array(device_wires))
-                basis_states = qml.math.convert_like(basis_states, state)
-                num = int(qml.math.dot(state, basis_states))
-
-                self._state = self._create_basis_state(num)
-
-            def apply_lightning(self, state, operations):
-                """Apply a list of operations to the state tensor.
-
-                Args:
-                    state (array[complex]): the input state tensor
-                    operations (list[~pennylane.operation.Operation]): operations to apply
-                    dtype (type): Type of numpy ``complex`` to be used. Can be important
-                    to specify for large systems for memory allocation purposes.
-
-                Returns:
-                    array[complex]: the output state tensor
-                """
-                state_vector = np.ravel(state)
-
-                if self.use_csingle:
-                    # use_csingle
-                    sim = StateVectorC64(state_vector)
-                else:
-                    # self.C_DTYPE is np.complex128 by default
-                    sim = StateVectorC128(state_vector)
-
-                # Skip over identity operations instead of performing
-                # matrix multiplication with the identity.
-                skipped_ops = ["Identity"]
-
-                for o in operations:
-                    if o.name in skipped_ops:
-                        continue
-                    method = getattr(sim, o.name, None)
-
-                    wires = self.wires.indices(o.wires)
-                    if method is None:
-                        # Inverse can be set to False since qml.matrix(o) is already in inverted form
-                        method = getattr(sim, "applyMatrix")
-                        try:
-                            method(qml.matrix(o), wires, False)
-                        except AttributeError:  # pragma: no cover
-                            # To support older versions of PL
-                            method(o.matrix, wires, False)
-                    else:
-                        inv = False
-                        param = o.parameters
-                        method(wires, inv, param)
-
-                return np.reshape(state_vector, state.shape)
-
-            def apply(self, operations, rotations=None, **kwargs):
-                # State preparation is currently done in Python
-                if operations:  # make sure operations[0] exists
-                    if isinstance(operations[0], QubitStateVector):
-                        self._apply_state_vector(
-                            operations[0].parameters[0].copy(), operations[0].wires
-                        )
-                        del operations[0]
-                    elif isinstance(operations[0], BasisState):
-                        self._apply_basis_state(operations[0].parameters[0], operations[0].wires)
-                        del operations[0]
-
-                for operation in operations:
-                    if isinstance(operation, (QubitStateVector, BasisState)):
-                        raise DeviceError(
-                            "Operation {} cannot be used after other Operations have already been "
-                            "applied on a {} device.".format(operation.name, self.short_name)
-                        )
-
-                if operations:
-                    self._pre_rotated_state = self.apply_lightning(self._state, operations)
-                else:
-                    self._pre_rotated_state = self._state
-
-                if rotations:
-                    self._state = self.apply_lightning(np.copy(self._pre_rotated_state), rotations)
-                else:
-                    self._state = self._pre_rotated_state
 
 else:  # binaries not available
